requirements.txt
Django==3.2.3
djangorestframework==3.13.1
openai==0.8.0
requests==2.25.1
